Split devtools implementation into sensible modules.
